listar cursos por email arreglado

+++ GestionCursosGrupo3/src/main/java/com/gestioncursos/service/CursosService.java
@@ -21,7 +21,7 @@ public interface CursosService {
 	
 	public abstract CursosModel findCurso(int id);
 	
-	public abstract List<CursosModel> listAllCursosProfesor(int id);
+	public abstract List<CursosModel> listAllCursosProfesor(long id);
 //	public abstract List<Alumnos> listAllAlumnos(CursosModel cursosModel);
 	
 	public abstract List<CursosModel> listAllCursosByNivelAsc();
